+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Populate `resourceName` when adding an eHoldings resource to an agreement line via "eHoldings". (ERM-4007)
 * Update Tags label in "Titles" accordion of "Package" record. (UIEH-1488)
 * Show `createdDate` when `updatedDate` is not available in Access Status Types metadata. (UIEH-1490)
+* Remove requests with "undefined" UUID are sent when opening eHoldings Knowledge Base credentials settings. (UIEH-1512)
 
 ## [11.1.1] (https://github.com/folio-org/ui-eholdings/tree/v11.1.1) (2026-05-28)